fix(ggml-cuda): cast ncols to uint32_t in norm fast-path
Cast ncols to uint32_t to match mul_ncols type.
Prevents signed/unsigned comparison warnings.
No functional changes to logic or kernel behavior.

+++ ggml/src/ggml-cuda/norm.cu
@@ -451,7 +451,7 @@ static void rms_norm_mul_f32_cuda(const float *  x,
     // Lean fast-path: mul is a contiguous per-column weight broadcast across all
     // rows/channels/samples (the standard rms_norm weight), no add. Bit-identical to the
     // general kernel here but with far fewer params -> higher occupancy on the big block.
-    if (add == nullptr && mul_nrows == 1 && mul_nchannels == 1 && mul_nsamples == 1 && mul_ncols == ncols) {
+    if (add == nullptr && mul_nrows == 1 && mul_nchannels == 1 && mul_nsamples == 1 && mul_ncols == (uint32_t)ncols) {
         const int block_size = ncols < 1024 ? 256 : 1024;
         const dim3 block_dims(block_size, 1, 1);
         const size_t nbytes_shared = block_size > WARP_SIZE ? 32 * sizeof(float) : 0;
